Add methods version, links - self (HATEOAS), JWT expires to 1 day

+++ mercury/controllers/user.py
@@ -1,7 +1,7 @@
 # coding=utf-8
 
 from mercury.services import user as services_user
-from mercury.services.custom_exceptions import HTTPException
+from mercury.services.custom_exceptions import MethodVersionNotFound
 
 from flask import abort, request
 from flask_restful import Resource, marshal
@@ -21,21 +21,25 @@ class UserListAPI(Resource):
 
         :return: All users.
         """
-        if not services_user.check_if_user_is_admin(get_jwt_identity()):
-            abort(401)
-        return {'users': [marshal(user, services_user.user_fields) for user in services_user.select_users()]}
+        if request.headers.get('Accept-Version', '1.0') == '1.0':
+            if not services_user.check_if_user_is_admin(get_jwt_identity()):
+                abort(401)
+            return {'users': [marshal(user, services_user.user_fields) for user in services_user.select_users()]}
+        raise MethodVersionNotFound()
 
     def post(self):
         """POST
 
         :return: Persisted user's JSON or error.
         """
-        if not services_user.check_if_user_is_admin(get_jwt_identity()):
-            abort(401)
-        if not request.json:
-            abort(400)
-        user = {key: value for key, value in self.reqparse.parse_args().items() if value is not None}
-        return {'user': marshal(services_user.insert_user(user['username']), services_user.user_fields)}, 201
+        if request.headers.get('Accept-Version', '1.0') == '1.0':
+            if not services_user.check_if_user_is_admin(get_jwt_identity()):
+                abort(401)
+            if not request.json:
+                abort(400)
+            user = {key: value for key, value in self.reqparse.parse_args().items() if value is not None}
+            return {'user': marshal(services_user.insert_user(user['username']), services_user.user_fields)}, 201
+        raise MethodVersionNotFound()
 
 
 class UserAPI(Resource):
@@ -43,8 +47,10 @@ class UserAPI(Resource):
 
     def __init__(self):
         """UserAPI constructor."""
-        self.reqparse = services_user.get_request_parser()
-        super(UserAPI, self).__init__()
+        if request.headers.get('Accept-Version', '1.0') == '1.0':
+            self.reqparse = services_user.get_request_parser()
+            super(UserAPI, self).__init__()
+        raise MethodVersionNotFound()
 
     def get(self, id):
         """GET
@@ -52,9 +58,11 @@ class UserAPI(Resource):
         :param id: User's id to find.
         :return: User found as JSON.
         """
-        if get_jwt_identity() != id:
-            abort(401)
-        return {'user': marshal(services_user.select_user(id), services_user.user_fields)}
+        if request.headers.get('Accept-Version', '1.0') == '1.0':
+            if get_jwt_identity() != id:
+                abort(401)
+            return {'user': marshal(services_user.select_user(id), services_user.user_fields)}
+        raise MethodVersionNotFound()
 
     def put(self, id):
         """PUT
@@ -62,13 +70,15 @@ class UserAPI(Resource):
         :param id: User's id to find.
         :return: Persisted user's base informations as JSON or error.
         """
-        if get_jwt_identity() != id:
-            abort(401)
-        if not request.json:
-            abort(400)
-        user = services_user.select_user(id)
-        [user.__setattr__(key, value) for key, value in self.reqparse.parse_args().items() if value is not None]
-        return {'user': marshal(services_user.update_user(user), services_user.user_fields)}
+        if request.headers.get('Accept-Version', '1.0') == '1.0':
+            if get_jwt_identity() != id:
+                abort(401)
+            if not request.json:
+                abort(400)
+            user = services_user.select_user(id)
+            [user.__setattr__(key, value) for key, value in self.reqparse.parse_args().items() if value is not None]
+            return {'user': marshal(services_user.update_user(user), services_user.user_fields)}
+        raise MethodVersionNotFound()
 
     def delete(self, id):
         """DELETE
@@ -76,9 +86,11 @@ class UserAPI(Resource):
         :param id: User's id to find.
         :return: True if elimination was successful or False if elimination was not possible.
         """
-        if get_jwt_identity() != id:
-            abort(401)
-        return {'result': services_user.delete_user(id)}
+        if request.headers.get('Accept-Version', '1.0') == '1.0':
+            if get_jwt_identity() != id:
+                abort(401)
+            return {'result': services_user.delete_user(id)}
+        raise MethodVersionNotFound()
 
 
 class UserLoginAPI(Resource):
@@ -92,10 +104,9 @@ class UserLoginAPI(Resource):
 
         :return: User's access_token as JSON or error.
         """
-        if not request.json:
-            abort(400)
-        user = {key: value for key, value in self.reqparse.parse_args().items() if value is not None}
-        try:
+        if request.headers.get('Accept-Version', '1.0') == '1.0':
+            if not request.json:
+                abort(400)
+            user = {key: value for key, value in self.reqparse.parse_args().items() if value is not None}
             return {'user': marshal(services_user.login_user(user), services_user.user_login_fields)}
-        except HTTPException as ex:
-            return {'error': str(ex)}, ex.code
+        raise MethodVersionNotFound()
